requirements.txt
amqp==2.5.2
asgiref==3.2.7
Babel==2.8.0
billiard==3.6.3.0
braintree==3.59.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.2
certifi==2020.4.5
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-autotranslate==1.1.0
django-countries==6.1.2
django-heroku==0.3.1
django-localflavor==3.0.1
django-modeltranslation==0.14.4
django-parler==2.0.1
django-rosetta==0.9.4
flower==0.9.3
futures==3.1.1
gettext-windows==1.0
goslate==1.5.1
gunicorn==20.0.4
html5lib==1.0.1
humanize==0.5.1
idna==2.9
kombu==4.6.8
Pillow==7.0.0
polib==1.1.0
psycopg2==2.8.5
pycparser==2.20
Pyphen==0.9.5
python-gettext==4.0
python-stdnum==1.13
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
tornado==5.1.1
urllib3==1.25.8
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.0.1
